fix(linter): Fix JSON schema to deny additional properties for plugins enum.
This ensures that VS Code and other editors will not accept unknown values for the plugins field.

I believe `string` was allowed initially because of the way jsPlugins were going to be implemented, but then the jsPlugins plans changed to use a separate config field? I'm not 100% sure, but that appears to be the case. Anyway, the schema was never updated after that pivot, and so this fixes the problem. We could probably remove the `any_of` to simplify things, but eh I just want to fix this for now.