fix(ast/estree): correct raw transfer deserializer for `FormalParameter` (#15302)
Correct the custom deserializer for `FormalParameter`. Was using the span offsets from `BindingRestElement`, when should be `FormalParameter`. `span` is the 1st field in both structs, so this fix makes no difference to the generated deserializer code, but it makes sure it'll remain correct if `span` field in either struct moves position later on.