feat(allocator): add `Address::from_ref` method (#15318)
We are generally using `Address::from_ptr` to get the `Address` of a reference, relying on `&T` dereferencing to `*const T`.

Add an explicit method `Address::from_ref` for this, and document its pitfalls (which are the same as `Address::from_ptr`).

Update all uses of `from_ptr` to `from_ref`, except those in `oxc_traverse`, where creating references is not possible due to aliasing rules.

Once that problem is solved, we can make `from_ptr` take a `NonNull<T>` instead of `*const T`, and make `Address`'s inner type a `NonZeroUsize`, which will give `Address` a niche, and reduce `Option<Address>` from 16 bytes to 8.