refactor(language_server): pass `serde_json::Value` to `WorkspaceWorker` configuration changes (#15279)
This PR split the logic for `WorkspaceOption` and his `Options`.
`WorkspaceOption.option` can now be any `serde_json::Value`. ATM, the `WorkspaceWorker` is still holding the `Option`.
The goal is, that both `ServerLinter` and `ServerFormatter` gets an on `new` a `serde_json::Value` and handles its own options parsing.

That is the first step for splitting the tools and move them to `--lsp` flag for CLI.

Goal of this PR Stack:
- `WorkspaceWorker` should not care about different tool options
- Every tool (currently `ServerFormatter` & `ServerLinter`) should consume `serde_json::Value`
- Building a unified pattern for creating tools beside the current (experimental) ones