refactor(language_server): always tell client the `textDocument/formatting` capability (#15398)
For now, the server will tell the client it supports formatting.
The client still needs to tell the server to format it with `textDocument/formatting`, which requires mostly a setup on the client side (`editor.defaultFormatter` for vscode).

Later, the passing Tool to the `Backend` will tell about the capability :)