refactor(language_server): handle `ServerFormatter` it own restart on `workspace/didChangeConfiguration` (#15428)
Now the `ServerFormatter` is aligned with the new `handle_configuration_change` from the `ServerLinter`.
The `diagnostic_reports` from `ToolRestartChanges` is not really needed, but I am fine for now because it's then always `None`.

The `WorkspaceWorker::did_change_configuration` is now generic enough to be a future `Vec` of Tools :)