perf(lexer): hint to compiler that EOF only happens once (#15512)
In `Lexer::read_next_token`, reaching EOF is very uncommon (because it only happens once at end of the file). Inform compiler of this using a `#[cold]` function, so it makes "not at EOF" the default path.

Small improvement on all the lexer and parser benchmarks (+1% on most lexer benchmarks).