perf(formatter): avoid unnecessary allocation for `BinaryLikeExpression` (#15467)
Previously, we split the binary expression into a few parts (at least two) and stored them in a `Vec`. In some situations, the `Vec` will be printed directly without using the `Vec` API. This means that storing them in a `Vec` is an unnecessary allocation. This PR eliminates this kind of allocation by introducing `format_flattened_logical_expression`, which won't store any part in a `Vec`.