chore(linter): Add just task for oxlint config schema generation (#15539)
This was being duplicated in multiple places, so I moved
`generate_json_schema` into Oxlintrc to avoid needing to keep those two
places in sync.

`just linter-schema-json` generates the configuration_schema.json so you
don't have to run the tests to regenerate it, which was a very confusing
behavior to figure out as a contributor.

However, this does mean this is now part of the public interface of the
create, so maybe it shouldn't be like this?