docs(linter): Add `search: false` frontmatter to generated linter docs. (#15537)
This prevents the docs from indexing generated pages for search. See
https://vitepress.dev/reference/default-theme-search#example-excluding-pages-from-search

We could alternatively do this in the vitepress config side of things
and explicitly block any results with these three names, up to the
maintainers.

Searching for the CLI would bring up the `generated` pages, so there
would be two results and it was a bit confusing:
<img width="932" height="196" alt="Screenshot 2025-11-09 at 6 18 01 PM"
src="https://github.com/user-attachments/assets/02302d27-1b80-49f3-8cdd-eab58870108c"
/>

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>