feat(oxfmt): use `prettier` directly and bundle `prettier` (#15544)
`@prettier/sync` is the synchronous version of `prettier`, which uses the `Worker` to hack it and make it synchronous.

Everything worked fine before we wanted to bundle them together. However, it hangs when they are bundled. I haven't found the reason why it hangs. Anyway, I don't think `@prettier/sync` is a good choice because it is not `real` synchronous, which sacrifices some performance to make it available.

Note: I am not familiar with NAPI, so I'm not sure whether the API is used correctly.