feat(linter/plugins): implement deprecated `SourceCode` tokens methods (#15645)
Implement the following deprecated `SourceCode` methods, related to tokens:

* `getTokenOrCommentBefore`
* `getTokenOrCommentAfter`
* `isSpaceBetweenTokens`

The 1st 2 are currently unimplemented, but can be implemented when we add tokens APIs.

`isSpaceBetweenTokens` implementation is flawed at present, in the same way as `isSpaceBetween` is, due to lack of tokens. But probably close enough to be useful in many cases. Added tests which demonstrate the flaw.