docs(linter): Add config docs for no-restricted-globals rule. (#15662)
Part of #14743.

Generated docs:

```md

## Configuration

This rule accepts a configuration object with the following properties:

### restrictedGlobals

type: `Record<string, string>`

default: `{}`

Objects in the format
`{ "name": "event", "message": "Use local parameter instead." }`, which define what globals
are restricted from use.
```

This is technically a positional argument and so should maybe be held back until we have a solution for documenting this kind of configuration option?