test(coverage/typescript): Add reviewed tsc diagnostics error codes (#15542)
Finally fixes #11582

This PR adds a script that generates a list of TSC error codes.

The codes are collected based on the TypeScript version managed by `just submodules`.
And the list is then committed to Git.

When the submodule version is updated, the script is also called and the list is regenerated.

If TSC introduces new diagnostic codes, the diff will show the added code, which we need to review.
Either leave it for the parser to report or ignore it if we cannot support it.