feat(oxlint): add enabled? column to --rules cli output (#15213)
addresses #11599 

Adds an Enabled? column by adding a separate method accessed by the CLI
when the --rules option is provided

Moves where we handle the the --rules request to after the ConfigStore
is available so we can get the list of rules that are enabled.

<img width="731" height="973" alt="image"
src="https://github.com/user-attachments/assets/dccb2b61-2af5-4949-b239-8975490513ed"
/>

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>