feat(allocator, ast): introduce `UnstableAddress` trait (#15700)
`Address::from_ptr` is hard to use correctly (as evidenced by https://github.com/rolldown/rolldown/pull/6961).

Introduce a new trait `UnstableAddress` which has a few more guardrails.

### Migration

We should swap all usage of `Address::from_ptr` for `UnstableAddress::unstable_address`.

When `Address::from_ptr` is being used correctly, it can easily be swapped out:

```diff
- let address = Address::from_ptr(node);
+ let address = node.unstable_address();
```

But if `Address::from_ptr` was being used incorrectly, migrating to `unstable_address` will reveal that problem with a compilation error.

`unstable_address` (as the name suggests) does not give the same guarantees about the stability of the `Address` it returns as `GetAddress::address` does. The name of the method makes that explicit.

More details in the doc comment in this PR.

### Migration plan

* #15701 migrates all usage of `Address::from_ptr` and `Address::from_ref` in Oxc's codebase to `unstable_address` (except in `oxc_traverse`, which has particular requirements around aliasing, and can't use `unstable_address`).
* #15702 removes `Address::from_ref`.
* Once this PR lands in an Oxc release, we can migrate Rolldown to `unstable_address`.
* In next Oxc release after that, we'll make `Address::from_ptr` an unsafe method, to discourage its use.
* At a future date, we can migrate `oxc_traverse` to using `NonNull` pointers instead of `*const` pointers, and get rid of `Address::from_ptr` entirely.