feat(minifier): disable removal of unnecessary `use strict` directives for DCE (#15691)
This is kind of a workaround. The target of this PR is to solve https://github.com/rolldown/rolldown/issues/6879 and https://github.com/rolldown/rolldown/issues/6880.

While the proper way to fix them is to use the proper `ModuleKind` (described in https://github.com/rolldown/rolldown/issues/7009), `ModuleKind::Unambiguous` doesn't work as expected and is blocked by https://github.com/oxc-project/oxc/issues/15541.

So for the meanwhile, this PR can solve those issues indirectly. We can revert this PR once the issues above are solved properly, but I guess it's also fine to keep it as I think there's less value to remove `use strict` directives in DCE.