chore(linter): Reword LSP fix message to distinguish from other option. (#15708)
The `Disable rule-name for this line` and `Disable rule-name for this
file` options look very similar at a glance, I think simply
distinguishing the option for the whole file vs. the line is a bit
better developer experience.

Currently it looks like this:
<img width="414" height="188" alt="Screenshot 2025-11-14 at 9 52 07 PM"
src="https://github.com/user-attachments/assets/f4b15b13-1773-43f4-92f1-a5f9a1a4bf3c"
/>

Initially, I didn't notice the difference and thought it was a bug where
I was getting an option from the ESLint extension and another from
oxlint :)

I'm marking this as a chore, but I guess it's technically a fix... or
something? :shrug: