fix(ast): include rest properties when using `get_binding_identifiers` (#15710)
previously, the rest param of an `ArrayPattern`​ inside a `BindingPattern` would never be visited, this meant it had to be manually visited inside the linter. It also meant that the minifier incorrectly transformed `{ const { a, ...b } = {}; b = {} }`​ to `{ let const { a, ...b } = {} }` which is a difference in runtime behvaiour (before is a `TypeError`​, after is not.