fix(formatter): measuring multiline text in `fits_text` is incorrect (#15762)
Regressed by #15372. The changes are made to match previous behavior. In `fits_text`, once the text width is multiple lines, we should check whether the current line width plus the width of the text to the first line break is greater than the allowed line width to take action, but previously, we didn't include `plus the width of the text to the first line break`.