fix(formatter): `CRLF` issue in the member chain (#15764)
* Fixes: #15559

Use `SourceText::lines_before` to check the line break, which has already handled `CRLF` and is well-tested.

Thought:
Whether we can remove all `\r` checks to avoid this problem, only counting `\n` as a line break. `CR` is very rare now, and we have a few places where we are just checking `\n`.