feat(allocator/vec): add `Vec::into_bump_slice` method (#15770)
Add an `into_bump_slice` method in `Vec` that directly delegates to the InnerVec's `into_bump_slice`. The reason we should add this method rather than using `InnerVec`'s `into_bump_slice` directly is that the method requires taking `self`, which doesn't work with `Deref`.

The method comments are copied from InnerVec's, and the example has been changed to make it work with the current Vec.