refactor(language_server): Add formatter and linter features (#15767)
Currently, there is 2 ways to start LS for oxc apps.

- `oxc_language_server`
- `oxlint --lsp`
- (`oxfmt --lsp` will be added soon)

Having duplicates are fine for now, but `ServerFormatter` (and `oxc_formatter`) should not bundled in `oxlint --lsp` case.

This PR adds `linter` and `formatter` features to exclude.
For `oxc_language_server`, both features are enabled by default, keep it as-is.