docs(linter): Improve diagnostic for consistent-type-definitions rule. (#15752)
Fix the grammar of the diagnostic message and improve the docs a bit.

This removes the redundant help message, but that results in the fix message showing up, which can look a bit weird for multi-line fixes? Not sure if I should keep this as-is or add a static help message.

Also adds an extra test case.