perf(formatter): reduce the size of `TextWidth` to 4 byte (#15827)
* close: #15823

The size of `TextWidth` was increased in https://github.com/oxc-project/oxc/pull/15762, which caused a 2% performance regression; I am unsure why this occurred at the time. Now I understand that the `TextWidth` size increased from 4 bytes to 8 bytes, which caused the FormatElement enum size to become 24 bytes in a 32-bit system.

This PR changes the `TextWidth` enum to a struct that only stores a `u32` width and uses the 31st bit as a marker to represent whether it is a multiline or not. Although it is smaller than `NonZeroU32`, it is enough to store the text width.