refactor(oxfmt): Use stdout and stderr appropreately (#15853)
Until now, everything was displayed to `stdout`.

This caused inconveniences when combining pipes in the CLI.
For example, the following does not work properly in some situation.

```sh
oxfmt dir_with_parse_error --list-different | wc -l
```

Because the graphical parse error flows along with the paths that has updated.

This PR solves this.