fix(linter): Fix `oxc/bad_array_method_on_arguments` rule behavior. (#15854)
This was allowing a number of methods which are entirely invalid to call on the `arguments` object, based on what TypeScript reports and from testing most of these additions in Firefox.

It was missing some methods that have been added in the last few years, including findLast, findLastIndex, toReversed, etc.

It also was still referring to `groupBy` as `groupToMap`.

Also improve the docs and all that.

I think we should consider adding a rule to just ban the usage of `arguments` entirely, with rest parameters being fully standardized for a decade now.

[TS Playground example](https://www.typescriptlang.org/play/?target=9#code/GYVwdgxgLglg9mABMMAKAlIg3gKEYgQwCcBzEAWwFMwoBnAOmBjABMAZA2qVDRAXgB82AL7oA3DmE4coSLATIwAJl658xMlRoMmrDlwCSrSgA8emQSPGTps6PCQoAzKryFSFanXokicEAAOAEIAnub8QliiElIy4PYKKAAsqu6aXgxQcABKlABulES0lCzhllGYUnbyjmAArKkantr0WQDKcERQJWWRoohV8TWKAGyNHlre7QEANjAQPQAMlUA)