test(linter/plugins): rename dirs in nested tests (#15888)
Rename the directories containing files to lint in the nested config tests from `nested` to `files`, to match the rest, and only lint the `files` directories.

I don't think this has any effect on the operation of the tests, but makes it easier to lint these directories as part of the repo (#15887).

This also removes the need for `overrideFiles` property in `testFixture` function's options, so remove it.