fix(oxlint/lsp): don't register `textDocument/formatting` capability (#15882)
Refactored the language server, so the `ServerCapability` can modified by each tool.
Before, we just registered the `textDocument/formatting` capability, if the client supports it dynamically.

Now we tell the client at the beginning if the request is supported, skipping the dynamic registration. Improving integration with client that does only support static registration.

And only of course when we use the formatter tool ;) Skipping it for `oxlint --lsp`