docs(linter): Add config docs generation for rules with Regex arguments (#15978)
~~Draft until we merge https://github.com/oxc-project/schemars/pull/9 and upgrade the oxc-schemars crate in this repo afterwards. Once we have the schemars crate bumped up to 0.8.27 we can safely rebase this branch and this'll be ready to merge.~~

But I can confirm that this now works with my patch to schemars :D

Part of #14743.

Generated docs:

```md
## Configuration

This rule accepts a configuration object with the following properties:

### ignore

type: `string[]`

A list of patterns to ignore when checking `catch` variable names. The pattern
can be a string or regular expression.

### name

type: `string`

default: `"error"`

The name to use for error variables in `catch` blocks. You can customize it
to something other than `'error'` (e.g., `'exception'`).
```

Unfortunately it doesn't recognize the default value in this case, although I'm not 100% sure why. But at the very least, this works and allows us to finish the docs for most of the remaining lint rules.

```md
## Configuration

This rule accepts a configuration object with the following properties:

### rejectPattern

type: `[
  string,
  null
]`

Regex pattern used to validate the `reject` parameter name. If provided, this pattern
is used instead of the default `^_?reject$` check.

### resolvePattern

type: `[
  string,
  null
]`

Regex pattern used to validate the `resolve` parameter name. If provided, this pattern
is used instead of the default `^_?resolve$` check.
```