refactor(linter/plugins): reduce repeat code initializing `tokensWithComments` (#16017)
The same code was repeated in many places. Move it into a function.

Function is only called when `tokensWithComments` is not already initialized (cold path), so no function call cost in majority of cases. In Rust, the pattern of moving code which is rarely called into a separate function is usually positive for perf. Not sure whether JS engine works the same way, but it might well be that V8 makes inlining decisions in much the same way.