test(ast/estree): bump `acorn-test262` submodule (#16108)
Bump `acorn-test262` submodule to latest. It aligns its version of `test262` and `TypeScript` submodules to same versions as used in this repo, which fixes a bunch of erroneous ESTree conformance test failures.

It also updates to latest version of TS-ESLint parser, which altered the shape of `TSImportType` (https://github.com/typescript-eslint/typescript-eslint/pull/11591). The test cases broken by this are fixed in next PR (#16109).