refactor(formatter): make `Format::fmt` non-failable (#16036)
This PR is a follow-up to https://github.com/astral-sh/ruff/pull/6613.

### Why can we do it

https://github.com/oxc-project/oxc/blob/622e000934d97ade5588f1e8cc3badde96e7328e/crates/oxc_formatter/src/formatter/diagnostics.rs#L11-L29

We can also make it non-failable since the entire program only has two points that can trigger an error, neither of which needs to interrupt the program immediately, and both can be improved to avoid making FormatResult everywhere

1. `FormatError::PoorLayout`

This error may occur only when handling call arguments, and only the first or the last argument is a `Function` or `ArrowFunctionExpression`. The error may be thrown during the formatting of them. I commented out the related logic of this in this PR and fixed it in #16093 to make it easier to review

2. `FormatError::InvalidDocument`

This is a printer error, but it may also occur only when we format `TaggedTemplateExpression`. There is a formatting logic that will format the template expression and print the IR immediately. If the IR we generated is incorrect, then the error will occur, but this is quite rare! So it's not worth using `FormatResult` everywhere to handle it. Currently, I just `unwrap` it. If we really need to handle it, we can store the error somewhere and return it after formatting in the future.

https://github.com/oxc-project/oxc/blob/78cd5910985cece594e74884717f77ad75347b8a/crates/oxc_formatter/src/write/template.rs#L602-L603

### Key changes

The most significant change is:

```diff
pub trait Format<'ast, T = ()> {
    - fn fmt(&self, f: &mut Formatter<'_, 'ast>) -> FormatResult<()>;
    + fn fmt(&self, f: &mut Formatter<'_, 'ast>);
}
```

All remaining changes to the order stem from the above change:

1. Remove all `FormatResult` usages
2. Remove all `?` operators that pair with the `format!` macro
4. Commented out `PoorLayout` related logic, which is going to be fixed in the #16093

### Benefits

We don't need to deal with `Option`, and removing all `Option` handling makes the code cleaner and more readable. Of course, the performance improved a little bit